# Improve Proxy Behaviour With the IR Optimizer